requirements.txt

fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
